Merge요청
[UPDATE] orderList.jsp에 쿼리스트림이 없을 때 띄우는 화면 설정, OrderDao.java에
DB로 부터 회원아이디와 주문번호에 해당하는 주문정보와 상품정보들을 반환하는 코드 작성
[ADD] orderDetail.jsp 주문상세정보를 조회하는 파일

+++ src/main/java/semi/dto/OrderItemDto.java
@@ -5,11 +5,11 @@ import java.util.Date;
 public class OrderItemDto {
 	// Order order
 	private int orderNo;
-	private int totalPrice;
+	private long totalPrice;
 	private String status;
 	private Date orderCreatedDate;
 	// OrderItem
-	private int orderProductPrice;
+	private long orderProductPrice;
 	private int orderProductQuantity;
 	// ProductItem
 	private int productItemNo;
@@ -30,11 +30,11 @@ public class OrderItemDto {
 		this.orderNo = orderNo;
 	}
 
-	public int getTotalPrice() {
+	public long getTotalPrice() {
 		return totalPrice;
 	}
 
-	public void setTotalPrice(int totalPrice) {
+	public void setTotalPrice(long totalPrice) {
 		this.totalPrice = totalPrice;
 	}
 
@@ -54,11 +54,11 @@ public class OrderItemDto {
 		this.orderCreatedDate = orderCreatedDate;
 	}
 
-	public int getOrderProductPrice() {
+	public long getOrderProductPrice() {
 		return orderProductPrice;
 	}
 
-	public void setOrderProductPrice(int orderProductPrice) {
+	public void setOrderProductPrice(long orderProductPrice) {
 		this.orderProductPrice = orderProductPrice;
 	}
